@@ -1,100 +1,46 @@
 [project]
 name = "marvin"
-dynamic = ["version"]
-description = "A lightweight AI engineering toolkit for building natural language interfaces that are reliable, scalable, and easy to trust."
+version = "3.0.0"
+description = ""
 readme = "README.md"
-license = { file = "LICENSE" }
-classifiers = [
-    "Programming Language :: Python :: 3 :: Only",
-    "Topic :: Scientific/Engineering :: Artificial Intelligence",
-]
-keywords = ["ai", "chatbot", "llm", "NLP", "natural language processing"]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
-    "cachetools>=5",
-    "fastapi",
-    "httpx>=0.24.1",
-    "jinja2>=3.1.2",
-    "jsonpatch>=1.33",
-    "openai>=1.21.0",
-    "prompt-toolkit>=3.0.33",
-    "pydantic>=2.4.2",
-    "pydantic_settings",
-    "rich>=12",
-    "tiktoken>=0.4.0",
-    "typer>=0.9.0",
-    "typing_extensions>=4.0.0",
-    "tzdata>=2023.3",           # need for windows
-    "uvicorn>=0.22.0",
-    "partialjson>=0.0.5",
+    "pydantic-settings>=2.6.1",
+    "pydantic[email]>=2.10.2",
+    "pydantic-ai>=0.0.20",
+    "aiosqlite>=0.20.0",
+    "sqlalchemy[asyncio]>=2.0.36",
+    "jinja2>=3.1.4",
+    "rich>=13.9.4",
+    "typer>=0.15.1",
 ]
 
 [project.optional-dependencies]
-generator = ["datamodel-code-generator>=0.20.0"]
-chromadb = ["chromadb"]
-prefect = ["prefect"]
-dev = [
-    "marvin[tests]",
-    "black[jupyter]",
-    "ipython",
-    "mkdocs-autolinks-plugin~=0.7",
-    "mkdocs-awesome-pages-plugin~=2.8",
-    "mkdocs-markdownextradata-plugin~=0.2",
-    "mkdocs-jupyter>=0.24.1",
-    "mkdocs-material[imaging]>=9.1.17",
-    "mkdocstrings[python]~=0.22",
-    "pdbpp~=0.10",
-    "pre-commit>=2.21,<4.0",
+test = [
+    "pre-commit",
+    "pytest-xdist>=3.6.1",
+    "pytest>=8.3.3",
+    "pytest-asyncio>=0.23.5",
     "ruff",
+    "pytest-timeout>=2.3.1",
+    "pytest-env>=1.1.5",
+    "pytest-rerunfailures>=15.0",
+    "chromadb>=0.6.0",
+    "dirty-equals>=0.9.0",
 ]
-tests = [
-    "pytest-asyncio>=0.18.2,!=0.22.0,<0.23.0",
-    "pytest-env>=0.8,<2.0",
-    "pytest-rerunfailures>=10,<14",
-    "pytest-sugar>=0.9,<2.0",
-    "pytest~=7.3.1",
-    "pytest-timeout",
-    "pytest-xdist",
-]
-audio = [
-    "SpeechRecognition>=3.10",
-    "PyAudio>=0.2.11",
-    "pydub>=0.25",
-    "simpleaudio>=1.0",
-]
-video = ["opencv-python >= 4.5"]
+dev = ["marvin[test]", "copychat>=0.5.2", "ipython>=8.12.3", "pdbpp>=0.10.3"]
 
-slackbot = [
-    "pydantic-ai[anthropic]@git+https://github.com/pydantic/pydantic-ai.git@main",
-    "prefect",
-    "numpy",
-    "raggy",
-    "turbopuffer==0.1.23",
-]
-
-[project.urls]
-Code = "https://github.com/prefecthq/marvin"
-Documentation = "https://www.askmarvin.ai"
-
+# Add this section for CLI entry point
 [project.scripts]
-marvin = "marvin.cli:app"
-
-[build-system]
-requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
-build-backend = "setuptools.build_meta"
-
-[tool.setuptools_scm]
-write_to = "src/marvin/_version.py"
+marvin = "marvin.cli.main:app"
 
 # pytest configuration
 [tool.pytest.ini_options]
-markers = [
-    "llm: indicates that a test calls an LLM (may be slow).",
-    "no_llm: indicates that a test does not require an LLM.",
-]
-timeout = 20
+addopts = "--verbose --color=yes"
+asyncio_mode = "auto"
+asyncio_default_fixture_loop_scope = "session"
+timeout = 60
 testpaths = ["tests"]
-
 norecursedirs = [
     "*.egg-info",
     ".git",
